+++ Sources/SwiftUIKit/Lists/ListButtonGroup.swift
@@ -3,7 +3,7 @@
 //  SwiftUIKit
 //
 //  Created by Daniel Saidi on 2024-03-15.
-//  Copyright © 2024-2025 Daniel Daniel Saidi. All rights reserved.
+//  Copyright © 2024-2025 Daniel Saidi. All rights reserved.
 //
 
 #if os(iOS)